server: move the drain code to its own file
Release note: None
